# Fix performance tests for LogManager and Run