--- a/Framework/API/src/LogManager.cpp
+++ b/Framework/API/src/LogManager.cpp
@@ -32,18 +32,6 @@ bool convertSingleValue(const Property *property, double &value) {
   }
 }
 
-/// Converts numeric property to double
-bool convertSingleValue(const Property *property, double &value) {
-  // The first one to succeed short-circuits and the value is returned.
-  // If all fail, returns false.
-  return convertSingleValue<double>(property, value) ||
-         convertSingleValue<int32_t>(property, value) ||
-         convertSingleValue<int64_t>(property, value) ||
-         convertSingleValue<float>(property, value) ||
-         convertSingleValue<uint32_t>(property, value) ||
-         convertSingleValue<uint64_t>(property, value);
-}
-
 /// Templated method to convert time series property to single double
 template <typename T>
 bool convertTimeSeriesToDouble(const Property *property, double &value,
@@ -80,23 +68,26 @@ bool convertTimeSeriesToDouble(const Property *property, double &value,
   }
 }
 
-/// Converts time series property to single double
-bool convertTimeSeriesToDouble(const Property *property, double &value,
-                               const Math::StatisticType &function) {
-  // Double first - can use filterByStatistic in this case
-  if (auto seriesDouble =
-          dynamic_cast<const TimeSeriesProperty<double> *>(property)) {
-    value = Mantid::Kernel::filterByStatistic(seriesDouble, function);
-    return true;
-  } else {
-    // Try other types, first to succeed short-circuits and returns.
-    // If none succeed, return false.
-    return convertTimeSeriesToDouble<int32_t>(property, value, function) ||
-           convertTimeSeriesToDouble<int64_t>(property, value, function) ||
-           convertTimeSeriesToDouble<float>(property, value, function) ||
-           convertTimeSeriesToDouble<uint32_t>(property, value, function) ||
-           convertTimeSeriesToDouble<uint64_t>(property, value, function);
-  }
+/// Templated method to convert a property to a single double
+template <typename T>
+bool convertPropertyToDouble(const Property *property, double &value,
+                             const Math::StatisticType &function) {
+  return convertSingleValue<T>(property, value) ||
+         convertTimeSeriesToDouble<T>(property, value, function);
+}
+
+/// Converts a property to a single double
+bool convertPropertyToDouble(const Property *property, double &value,
+                             const Math::StatisticType &function) {
+  // Order these with double and int first, and less likely options later.
+  // The first one to succeed short-circuits and the value is returned.
+  // If all fail, returns false.
+  return convertPropertyToDouble<double>(property, value, function) ||
+         convertPropertyToDouble<int32_t>(property, value, function) ||
+         convertPropertyToDouble<int64_t>(property, value, function) ||
+         convertPropertyToDouble<uint32_t>(property, value, function) ||
+         convertPropertyToDouble<uint64_t>(property, value, function) ||
+         convertPropertyToDouble<float>(property, value, function);
 }
 }
 
@@ -346,8 +337,7 @@ double LogManager::getPropertyAsSingleValue(
   const auto key = std::make_pair(name, statistic);
   if (!m_singleValueCache.getCache(key, singleValue)) {
     const Property *log = getProperty(name);
-    if (!convertSingleValue(log, singleValue) &&
-        !convertTimeSeriesToDouble(log, singleValue, statistic)) {
+    if (!convertPropertyToDouble(log, singleValue, statistic)) {
       if (const auto stringLog =
               dynamic_cast<const PropertyWithValue<std::string> *>(log)) {
         // Try to lexically cast string to a double