PR #17105, which made changes to ``LogManager::getPropertyAsSingleValue``, broke two performance tests (see issue). The reason for the worse performance is that the values were no longer cached and were obtained fresh on each call.

Fix the tests by putting the caching back. Also refactored the methods so that there should be fewer casts required to deal with the most common cases, but the performance improvement comes from the caching, not the refactoring.

**To test:**

- All unit tests should still pass - no change in functionality.
- Run the performance tests for `LogManager` and `Run`: see http://www.mantidproject.org/Writing_Performance_Tests#Running_Performance_Tests for how.
- Compare the test results at three points (run a few times for each):
  - Before PR #17105 was merged
  - After PR #17105 was merged
  - After the changes in the current PR

The results I get (Win7, Debug) are as follows:
**LogManagerTestPerformance**
*Original*           0.09s
*After #17105*   1.56 s
*Changes here*  0.09 s

**RunTestPerformance**
*Original*           0.09s
*After #17105*   1.55 s
*Changes here*  0.09 s

Fixes #17197.

*Does not need to be in the release notes* (test was broken more recently than the release)


---

#### Reviewer ####

Please comment on the following ([full description](http://www.mantidproject.org/Individual_Ticket_Testing)):

##### Code Review #####

- [x] Is the code of an acceptable quality?
- [x] Does the code conform to the [coding standards](http://www.mantidproject.org/Coding_Standards)? Is it well structured with small focussed classes/methods/functions?
- [ ] Are there unit/system tests in place? Are the unit tests small and test the a class in isolation?
- [ ] If there are changes in the release notes then do they describe the changes appropriately?

##### Functional Tests #####

- [x] Do changes function as described? Add comments below that describe the tests performed?
- [ ] How do the changes handle unexpected situations, e.g. bad input?
- [ ] Has the relevant documentation been added/updated?
- [ ] Is user-facing documentation written in a user-friendly manner?
- [ ] Has developer documentation been updated if required?

- Does everything look good? Comment with the ship it emoji but don't merge. A member of `@mantidproject/gatekeepers` will take care of it.