@tom-perkins 2016-08-15
Merged in `master` to include the fix for failing system test