@DavidFair 2016-09-08
Windows 10 - Release:
**LogManagerTestPerformance**
Before #17105: 0.002 s
After #17105: 0.02 s
After PR: 0.002 s
**RunTestPerformance**
Before #17105: 0.002 s
After #17105: 0.02 s
After PR: 0.002s

Everything looks good from my end - lets :shipit: